Update relationships-in-standard-dataflows.md
Good summary of this topic, I think it needs a bit more volume to account for all audience levels.


1. In the "## What are Lookup or Relationship fields?" section, for example, I would leverage content and link from the CDS article about it: https://docs.microsoft.com/en-us/powerapps/maker/common-data-service/data-platform-entity-lookup

You can also grab some content from it, for example I find this or some of this passage useful:
"Data in one entity often relates to data in another entity. For example, you might have a Teachers entity and a Class entity, and the Class entity might have a lookup relation to the Teachers entity to show which teacher teaches the class. You can use a lookup field to show data from the Teachers entity. This is commonly referred to as a lookup field."

Read more about relationships and lookupfields .... [ Link to above article]

I know you created your own example - which is okay, I would use region etc. in this doc, but point is that first define a bit the relationship in words/real life example and then re-leverage throughout the doc. Without concrete names, it makes it harder to grasp for someone that just started with lookups/relationships. 

2. For example, in this paragraph:
"If you have created a lookup field in one entity that points to another entity, that field in the mapping of the dataflow, might not appear. The reason is that each entity requires a Key field. This best practice would then makes sure that the field is mappable in the entity mappings of the dataflow."

It would be helpful to leverage the above example:
If you just created a teacher lookup field in the class entity that points to the teacher entity, (Diagram would be useful here too) the teacher lookup field might not appear in the dataflow's map entities experience. The reason is that both class and teacher entities requires a Key field. This best practice would then makes sure that the field is mappable in the entity mappings of the dataflow.