Update salesforce-objects.md
Many of our customers are suffering from this issue when they use the lightning URL to connect Salesforce object. I know that we have write this limitation in our doc, but it seems not obvious for cx. so, I suggest emphasizing this limitation. BTW, I just get some feedbacks, that usually customer will only get the Lightning URL from their admin. So, it would be better if we can consider support the lightning URL.