Update dataverse.md
The documentation incorrectly mentions that the connector has 80MB size limit. There is no limit on the size of the data rather the limit is on the time-out. refere to this documentation https://learn.microsoft.com/en-us/power-apps/developer/data-platform/dataverse-sql-query#limitations

Also confirm this with Scott Sewell, PM Fabric CAT.