Update azure-analysis-services.md
Changes performed as suggested by Doug on comment "@RESET-USER - The capabilities section is supposed to list the capabilities that can be directly set from the initial dialog box for the connector (that is, the dialog box that comes up when you select the connector in the get data experience). It may be true that the Azure Analysis Services connector supports DirectQuery, but it isn't through the initial connection. If DirectQuery is supported through a different setting, we need to add a section to the connector that explains how to use that setting (or at least explain where the setting is used and then include a link to the section of the article that you supply above)."